refactor: рефактор страницы AboutUs
