+++ src/pages/AboutUs/index.js
@@ -1,11 +1,6 @@
 import BasePage from '../../layout/BasePage';
 import CardFigure from '../../components/Cards/CardFigure/CardFigure';
 import CardAbout from '../../components/Cards/CardAbout/CardAbout';
-import {
-  TitleH3,
-  Blockquote,
-  LogoBlue,
-  Card,
-} from '../../components/utils/index';
+import { Blockquote, Card, LogoBlue, TitleH3 } from '../../components/utils';
 
 export { BasePage, TitleH3, Blockquote, LogoBlue, CardFigure, CardAbout, Card };
